% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drawBackground}
\alias{drawBackground}
\title{Draw the background.}
\usage{
drawBackground(title, date, lang = "zh")
}
\arguments{
\item{title}{the image's title}

\item{date}{the date}

\item{lang}{the language zh or en}
}
\description{
Draw the background.
}

