% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drawDescription}
\alias{drawDescription}
\title{Draw the description.}
\usage{
drawDescription(data, temp, lang = "zh")
}
\arguments{
\item{data}{daily data}

\item{temp}{the temperature}

\item{lang}{the language zh or en}
}
\description{
Draw the description.
}

