% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drawTemperature}
\alias{drawTemperature}
\title{Draw temperature picture.}
\usage{
drawTemperature(data, lang = "zh", type = "high", date = Sys.time(),
  output = FALSE, path = "")
}
\arguments{
\item{data}{daily data}

\item{lang}{language}

\item{type}{low or high}

\item{date}{the date}

\item{output}{output a file or not}

\item{path}{image output position}
}
\description{
Draw temperature picture.
}

