% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filename}
\alias{filename}
\title{Define a filename from current date.}
\usage{
filename(date = Sys.time())
}
\arguments{
\item{date}{input a date type}
}
\value{
character a file name
}
\description{
Define a filename from current date.
}
\examples{
\dontrun{
 filename()
 filename(as.Date("20110701",format="\%Y\%m\%d"))
}
}
\keyword{filename}

