% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCityInfo}
\alias{getCityInfo}
\title{Get all of city info}
\usage{
getCityInfo(lang = "en")
}
\arguments{
\item{lang}{input a language}
}
\value{
data.frame city info
}
\description{
Get all of city info
}
\examples{
\dontrun{
 getCityInfo()
 getCityInfo(lang="en")
 getCityInfo(lang="zh")
}
}
\keyword{language}

