% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWeatherByCity}
\alias{getWeatherByCity}
\title{Get one city weather Data.}
\usage{
getWeatherByCity(en = "beijing", src = "yahoo")
}
\arguments{
\item{en}{input a English city name}

\item{src}{input data source}
}
\value{
data.frame weather data
}
\description{
Get one city weather Data.
}
\examples{
\dontrun{
 getWeatherByCity()
 getWeatherByCity(en="beijing")
}
}
\keyword{weather}

