% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWeatherFromYahoo}
\alias{getWeatherFromYahoo}
\title{Get weather data from Yahoo openAPI.}
\usage{
getWeatherFromYahoo(woeid = 2151330)
}
\arguments{
\item{woeid}{input a yahoo woeid}
}
\value{
data.frame weather data
}
\description{
Get weather data from Yahoo openAPI.
}
\examples{
\dontrun{
 getWeatherFromYahoo()
 getWeatherFromYahoo(2151330)
}
}
\keyword{weather}

