% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{weather20141001}
\alias{weather20141001}
\title{Dataset for 20141001.}
\format{A data frame}
\source{
\url{http://weather.yahooapis.com/forecastrss?w=2151330}
}
\usage{
weather20141001
}
\description{
A demo dataset.
}
\keyword{datasets}

