% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{WOEID}
\alias{WOEID}
\title{The yahoo code for weather openAPI.}
\format{A data frame}
\source{
\url{https://developer.yahoo.com/geo/geoplanet/guide/concepts.html}
}
\usage{
WOEID
}
\description{
The yahoo code for weather openAPI.
}
\keyword{datasets}

